requirements.txt
python-dotenv
python-telegram-bot
backtrader
ccxt
git+https://github.com/wilfredbtan/bt-ccxt-store.git
xlsxwriter
pandas
termcolor